refactor(cli): split config.rs by the table each section describes
1,576 production lines holding five unrelated `[table]`s and the loader for all
of them. Split so a section is where its name says it is:

- `config/policy.rs` (76) - `ToolPolicy`, `ScriptPermission`,
  `ScriptToolPermissions`. Referred to by every other section, so they sit apart
  rather than inside whichever one happened to need them first.
- `config/security.rs` (218) - `SecurityConfig`, `ReadPathGrants`. The shipped
  posture: what a blueprint may loosen, what a shell sees, where `[read_paths]`
  applies.
- `config/limits.rs` (354) - `LimitsConfig`, `WebhookConfig` and their serde
  defaults, kept beside the fields they fill.
- `config/providers.rs` (117) - `ProviderConfig` (with its hand-written `Debug`,
  which is hand-written so an API key cannot reach a log through a derive) and
  `ModelProviderConfig`.
- `config/mod.rs` - `Config` itself, its loader, and the filesystem helpers.

Glob re-exported per `pipeline/mod.rs`, so every existing `config::SecurityConfig`
path is unchanged and no call site moved.

**Verified pure by sorted-line diff: 2,366 in, 2,366 out.** The entire
difference is two functions gaining `pub(crate)`, and that is forced rather than
incidental: serde resolves a `default = "..."` path in the module the struct is
defined in, so `default_true` - used by three sections - has to be reachable
from all three. It now lives beside `Config` with a comment saying why.

Two traps worth recording, both from moving a file one directory deeper:

- `include_str!` resolves relative to the **file**, so the two schema paths in
  the tests needed another `../`. This is the same trap the manifest split hit.
- A serde default referenced from a sibling module has to be `pub(crate)`;
  the glob re-export does not carry private items.

213 config tests pass and the package holds 100%.
