+++ crates/leviath-cli/src/config/mod.rs
@@ -6,81 +6,17 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::PathBuf;
 
-/// Whether a tool call should execute automatically or require user approval.
-///
-/// The effective policy for a tool is resolved by narrowest scope first:
-/// launch-flag > stage > agent > global config > built-in default.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum ToolPolicy {
-    /// Execute without prompting.
-    Allow,
-    /// Ask the user before each call (or once per session with `allow_session`).
-    #[default]
-    Ask,
-    /// Never execute - return a denied error to the model.
-    Deny,
-}
-
-// `TitleConfig` (plain data used by the engine's title generation) lives in
-// `leviath_core::config` so `leviath-runtime` can reference it without a CLI
-// dependency. Re-exported here so `crate::config::TitleConfig` paths resolve.
-pub use leviath_core::config::TitleConfig;
-
-// Same arrangement for the `[observability]` section: the plain data lives in
-// `leviath_core::config` (the telemetry sink crate reads it), re-exported here.
-pub use leviath_core::config::{ObservabilityConfig, TelemetryExporterKind};
-
-/// Permission for one Rhai *script-tool* host function (Layer 3 of the
-/// four-layer permission model). Gates what a registered script may *do*,
-/// independent of
-/// whether the tool itself is visible ([`available_tools`]) or approved at
-/// runtime ([`ToolPolicy`]).
-///
-/// [`available_tools`]: leviath_core::blueprint::Stage::available_tools
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum ScriptPermission {
-    /// The host function may run.
-    Allow,
-    /// The host function is blocked - the call returns a `[denied]` error.
-    Deny,
-    /// Defer to the agent's own `tool_permissions` for the equivalent built-in
-    /// (`read_file`/`shell`): permitted only when that resolves to
-    /// [`ToolPolicy::Allow`]. For the network/env functions (`http_get`,
-    /// `http_post`, `env_var`), which have no built-in equivalent, `Inherit`
-    /// permits the call (they're needed for tools to be useful, and the tool
-    /// itself is still gated by Layers 1/2/4).
-    #[default]
-    Inherit,
-}
-
-/// Per-host-function permissions for Rhai script tools (`[tool_script_permissions]`).
-///
-/// Every field defaults to [`ScriptPermission::Inherit`], so an unconfigured
-/// install lets network/env functions run while file/shell functions defer to
-/// the agent's own tool permissions.
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ScriptToolPermissions {
-    /// Permission for `http_get`.
-    #[serde(default)]
-    pub http_get: ScriptPermission,
-    /// Permission for `http_post`.
-    #[serde(default)]
-    pub http_post: ScriptPermission,
-    /// Permission for `shell`.
-    #[serde(default)]
-    pub shell: ScriptPermission,
-    /// Permission for `read_file`.
-    #[serde(default)]
-    pub read_file: ScriptPermission,
-    /// Permission for `write_file`.
-    #[serde(default)]
-    pub write_file: ScriptPermission,
-    /// Permission for `env_var`.
-    #[serde(default)]
-    pub env_var: ScriptPermission,
-}
+// Sections of the former single-file config, one per `[table]` it describes.
+// Glob re-exported so every existing `config::SecurityConfig` path keeps
+// working and the split stays a pure move.
+mod limits;
+pub use limits::*;
+mod policy;
+pub use policy::*;
+mod providers;
+pub use providers::*;
+mod security;
+pub use security::*;
 
 /// CLI configuration.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -308,695 +244,6 @@ pub struct Config {
     /// takes effect for a path the blueprint also declares.
     #[serde(default)]
     pub agent_read_paths: HashMap<String, ReadPathGrants>,
-}
-
-/// One agent's entry in `[agent_read_paths.<name>]`.
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ReadPathGrants {
-    /// Granted entries, same forms as a blueprint's `[read_paths] allow`.
-    #[serde(default)]
-    pub allow: Vec<String>,
-}
-
-/// `[security]` in `~/.leviath/config.toml`.
-///
-/// Distinct from a *blueprint's* `[security]` block, which configures taint
-/// tracking for one agent - this one holds machine-wide switches.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct SecurityConfig {
-    /// Directories a run's workdir may sit under without being confirmed.
-    ///
-    /// **Empty by default**, which means "ask once about anything alarming"
-    /// rather than "ask about everything": a workdir is only questioned when it
-    /// is a home directory or a filesystem root, which is where an agent's file
-    /// writes do the most damage and the least good. Listing a path here says
-    /// "yes, I work there" and silences the prompt for it and everything under
-    /// it.
-    ///
-    /// This exists because the workdir defaults to wherever `lev run` was
-    /// invoked, and running from `~` is an easy thing to do by accident - issue
-    /// #252 is a machine that lost 115 GB to an agent writing under a profile
-    /// root. Confirming is cheap; noticing afterwards is not.
-    #[serde(default)]
-    pub allowed_workdirs: Vec<String>,
-
-    /// Whether a blueprint's `seed = { command = "..." }` regions may run.
-    ///
-    /// **On by default.** A command seed executes at spawn - before the first
-    /// inference, and therefore before any tool-approval prompt - so it is the
-    /// one place a manifest can run something without the user being asked.
-    /// It is still confined to the run's workdir, routed through the entry
-    /// stage's sandbox when the agent declares one, and capped by
-    /// `[limits] script_shell_timeout_secs`. Set this to `false` to refuse them
-    /// machine-wide, or pass `--no-seed-commands` for a single run. Inspect an
-    /// agent's command seeds before installing it with `lev validate <path>`.
-    #[serde(default = "default_true")]
-    pub allow_seed_commands: bool,
-
-    /// Whether agent-driven fetches may reach loopback, private, and link-local
-    /// addresses.
-    ///
-    /// **Off by default.** An agent's `web_fetch` URL is chosen by the model out
-    /// of context an attacker can influence - a search result, a page fetched a
-    /// moment ago, an issue body - so an unrestricted fetch makes the agent a
-    /// confused deputy *inside* the user's network. The concrete targets are
-    /// `http://169.254.169.254/…` (cloud metadata, which returns instance
-    /// credentials), `http://127.0.0.1:3000/api/…` (the user's own `lev serve`),
-    /// and anything on the LAN.
-    ///
-    /// Turn this on when the agent is genuinely meant to talk to something local -
-    /// a self-hosted model, a dev server under test. It applies to the script
-    /// host's `http_get`/`http_post` and to redirect following; see
-    /// [`leviath_core::net`].
-    #[serde(default)]
-    pub allow_local_network: bool,
-
-    /// Credential-shaped environment variables that agent scripts may read.
-    ///
-    /// A Rhai script tool or script provider calling `env_var("NAME")` gets any
-    /// ordinary variable - `PATH`, `TZ`, an app's own config. A name that *looks
-    /// like a credential* (see [`leviath_core::secrets::is_sensitive_env_name`])
-    /// is refused unless it appears here, because a two-line script tool reading
-    /// `ANTHROPIC_API_KEY` and POSTing it elsewhere was otherwise a working
-    /// exfiltration path with no prompt anywhere in it.
-    ///
-    /// List the exact names a script legitimately needs - typically the key for
-    /// a custom provider script:
-    ///
-    /// ```toml
-    /// [security]
-    /// allow_env_vars = ["MY_PROVIDER_KEY"]
-    /// ```
-    ///
-    /// Matching is case-insensitive and exact. There is no wildcard: `"*"` is
-    /// read as a variable literally named `*`, not as "allow everything".
-    #[serde(default)]
-    pub allow_env_vars: Vec<String>,
-
-    /// How much of the daemon's environment a `shell` tool call, a Rhai
-    /// `shell()` host call, and a region's command seed inherit.
-    ///
-    /// The daemon holds provider keys, `LEVIATH_API_TOKEN`, and whatever
-    /// credentials the person who started it had exported. Handing all of that
-    /// to every shell command means a single `env` in tool output leaks the lot.
-    ///
-    /// `filtered` (the default) withholds credential-shaped names but keeps
-    /// `SSH_AUTH_SOCK`, so `git push` over agent keys still works. `strict`
-    /// drops the carve-out and also takes `AWS_PROFILE`, `KUBECONFIG` and
-    /// friends. `custom` ignores the shape heuristic and withholds exactly what
-    /// [`Self::shell_env_withhold`] names. `inherit` is the old behaviour.
-    ///
-    /// Toolchain variables - `PATH`, `HOME`, `CARGO_HOME`, `JAVA_HOME`,
-    /// `VIRTUAL_ENV`, `NVM_DIR`, `GOPATH`, `DOCKER_HOST` - pass through under
-    /// every mode. [`Self::allow_env_vars`] hands a specific name over under
-    /// every mode too.
-    #[serde(default)]
-    pub shell_env: leviath_core::ShellEnvMode,
-
-    /// The names `shell_env = "custom"` withholds. Ignored under every other
-    /// mode, where the name-shape heuristic decides instead.
-    #[serde(default)]
-    pub shell_env_withhold: Vec<String>,
-
-    /// Whether a blueprint's `[read_paths]` declarations are honored as-is.
-    ///
-    /// **Off by default.** A `[read_paths]` block travels inside the
-    /// `agent.leviath` you installed, and a manifest may only *tighten* what
-    /// your config allows, never widen it - otherwise any agent package could
-    /// read `~/.ssh`, this very config file (your API keys), or a password
-    /// store by shipping one TOML line. With this off, an agent's declared
-    /// read paths are inert until you grant them via [`Self::read_paths`] or
-    /// `[agent_read_paths.<name>]`. Turning it on says "any blueprint I run
-    /// may read every path it declares" - reads only, each access still
-    /// resolves symlinks and must land inside a declared entry, but prefer
-    /// the per-agent grant for anything you did not author yourself.
-    #[serde(default)]
-    pub allow_blueprint_read_paths: bool,
-
-    /// Honour every blueprint's own `[safe_commands]` block.
-    ///
-    /// **Off by default**, and for the same reason as
-    /// [`Self::allow_blueprint_read_paths`]: a `[safe_commands]` block travels
-    /// inside an `agent.leviath` you installed, so letting it count by itself
-    /// would let any agent package pre-approve its own shell with one TOML line.
-    /// With this off, a blueprint's list is inert until you opt in, either here
-    /// for every agent or per agent via
-    /// `[agent_safe_commands.<name>] allow_blueprint = true`. Prefer the
-    /// per-agent grant for anything you did not author yourself.
-    #[serde(default)]
-    pub allow_blueprint_safe_commands: bool,
-
-    /// Honour a blueprint's `[tool_permissions]` even where it is *more*
-    /// permissive than the built-in default for a tool you have not configured.
-    ///
-    /// Off by default, for the same reason as the two switches around it:
-    /// declaring is not granting. Saying nothing about `shell` is the normal
-    /// state, so without this a downloaded manifest could give itself
-    /// `shell = "allow"` on a stock machine. With it off, a blueprint may still
-    /// pre-approve the read-only web tools that are some agents' whole point,
-    /// and anything beyond that is clamped to the built-in default.
-    ///
-    /// The per-agent grant needs no switch of its own: naming the tool under
-    /// `[agent_tool_permissions.<name>]` makes it a ceiling for that agent, and
-    /// a blueprint may go up to a ceiling. Prefer that for anything you did not
-    /// author yourself - it says which agent and which tool, where this says
-    /// "every agent, every tool".
-    #[serde(default)]
-    pub allow_blueprint_permissions: bool,
-
-    /// Machine-wide read grants for agents that declare `[read_paths]`.
-    ///
-    /// Entries use the same three forms as a blueprint's `[read_paths] allow`:
-    /// an exact path (grants its subtree), `glob:` and `regex:` patterns
-    /// (matched against the symlink-resolved real path, written with `/` on
-    /// every OS, regex auto-anchored). `~/` expands to your home; a relative
-    /// entry resolves against the run's workdir.
-    ///
-    /// ```toml
-    /// [security]
-    /// read_paths = ["~/.leviath/runs", "glob:~/design-docs/**"]
-    /// ```
-    ///
-    /// A grant only takes effect for a path the running blueprint *also*
-    /// declares - by itself it grants nothing, so listing a directory here
-    /// does not open it to agents that never asked.
-    #[serde(default)]
-    pub read_paths: Vec<String>,
-
-    /// Where provider API keys and MCP OAuth tokens are kept.
-    ///
-    /// **`file` by default** - `~/.leviath/config.toml` and
-    /// `~/.leviath/mcp-auth.json`, both created `0600` so they are never even
-    /// briefly world-readable. This is what Claude Code and Codex do, and it is
-    /// the only backend that works headless, in a container, over SSH, and on a
-    /// CI runner.
-    ///
-    /// Set it to `keychain` to move secrets into the OS credential store (macOS
-    /// Keychain, Windows Credential Manager, Secret Service elsewhere), so a
-    /// stolen `~/.leviath` directory yields nothing:
-    ///
-    /// ```toml
-    /// [security]
-    /// credential_store = "keychain"
-    /// ```
-    ///
-    /// Then run `lev auth migrate` to move the secrets you already have. It is
-    /// opt-in rather than the default because an unavailable keychain is not a
-    /// degraded experience but a broken one - every inference fails at once -
-    /// and the environments Leviath is most useful in are the least likely to
-    /// have a working credential store. `lev auth status` reports whether this
-    /// machine actually has one.
-    #[serde(default)]
-    pub credential_store: leviath_core::CredentialStoreKind,
-}
-
-impl Default for SecurityConfig {
-    fn default() -> Self {
-        Self {
-            allowed_workdirs: Vec::new(),
-            allow_seed_commands: true,
-            allow_local_network: false,
-            allow_env_vars: Vec::new(),
-            shell_env: leviath_core::ShellEnvMode::default(),
-            shell_env_withhold: Vec::new(),
-            allow_blueprint_read_paths: false,
-            allow_blueprint_safe_commands: false,
-            allow_blueprint_permissions: false,
-            read_paths: Vec::new(),
-            credential_store: leviath_core::CredentialStoreKind::File,
-        }
-    }
-}
-
-fn default_true() -> bool {
-    true
-}
-
-fn default_max_concurrent_inferences() -> Option<usize> {
-    Some(8)
-}
-
-fn default_default_max_iterations() -> Option<usize> {
-    Some(50)
-}
-
-fn default_max_concurrent_tools() -> usize {
-    8
-}
-
-fn default_script_shell_timeout_secs() -> u64 {
-    60
-}
-
-fn default_stall_timeout_secs() -> u64 {
-    leviath_runtime::pipeline::DEFAULT_STALL_TIMEOUT_SECS
-}
-
-fn default_dead_cycles_before_relief() -> u32 {
-    leviath_runtime::host::DEFAULT_DEAD_CYCLES_BEFORE_RELIEF
-}
-
-fn default_mcp_idle_disconnect_secs() -> u64 {
-    crate::daemon::mcp_pool::DEFAULT_MCP_IDLE_DISCONNECT_SECS
-}
-
-fn default_finished_retention_secs() -> u64 {
-    leviath_runtime::host::DEFAULT_FINISHED_RETENTION_SECS
-}
-
-fn default_wedge_timeout_secs() -> u64 {
-    leviath_runtime::pipeline::DEFAULT_WEDGE_TIMEOUT_SECS
-}
-
-/// The provider a config that names none is assumed to mean.
-///
-/// Exists so `default_provider` can carry `#[serde(default)]`: without one,
-/// every field on [`Config`] that lacked a default made a hand-written
-/// `config.toml` a parse error. Writing three lines to point Leviath at
-/// OpenRouter used to fail with `missing field `providers``, which names a
-/// table the user has no reason to know about and says nothing about what to
-/// add. Kept in sync with [`Config::default`] by
-/// `an_empty_config_file_parses_to_the_defaults`.
-fn default_provider_name() -> String {
-    "anthropic".to_string()
-}
-
-fn default_provider_failures_before_open() -> u32 {
-    leviath_runtime::pipeline::DEFAULT_FAILURES_BEFORE_OPEN
-}
-
-fn default_provider_circuit_cooldown_secs() -> u64 {
-    leviath_runtime::pipeline::DEFAULT_CIRCUIT_COOLDOWN_SECS
-}
-
-fn default_interaction_timeout_secs() -> u64 {
-    leviath_runtime::interaction_hub::DEFAULT_INTERACTION_TIMEOUT_SECS
-}
-
-/// Runtime resource limits with safe defaults baked in.
-///
-/// Both fields default to a bounded value so a fresh install can't accidentally
-/// run unbounded inference concurrency or an unbounded agent loop. Set a field
-/// explicitly in `[limits]` to raise or lower it.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LimitsConfig {
-    /// Global fallback cap on concurrent inference requests for any model
-    /// without its own per-model pool entry. Defaults to `Some(8)`; omit or set
-    /// a large number to effectively unbound it.
-    ///
-    /// One physical bound sits behind this for *script* providers: each of
-    /// their in-flight calls occupies a blocking-pool thread, and the daemon's
-    /// runtime provisions 2048 of those. Pools above 2048 only run that wide
-    /// for HTTP providers, whose calls are fully async.
-    #[serde(default = "default_max_concurrent_inferences")]
-    pub max_concurrent_inferences: Option<usize>,
-
-    /// Size of the shared tool-execution worker pool - the number of agents whose
-    /// tool batches may run concurrently across the whole daemon (the tool-lane
-    /// counterpart of `max_concurrent_inferences`). Defaults to `8`. Clamped to at
-    /// least 1.
-    #[serde(default = "default_max_concurrent_tools")]
-    pub max_concurrent_tools: usize,
-
-    /// Fallback `max_iterations` applied to a stage that does not set its own,
-    /// so an agent can't loop forever with no completion signal. Defaults to
-    /// `Some(50)`. A stage's explicit `max_iterations` always wins.
-    #[serde(default = "default_default_max_iterations")]
-    pub default_max_iterations: Option<usize>,
-
-    /// Opt-in exact pre-inference token budgeting. When `true`, each agent
-    /// inference is preceded by an exact token count of the assembled request
-    /// (via the provider's `count_tokens`, which uses a remote endpoint for
-    /// Anthropic/Gemini and a local heuristic otherwise) and is rejected before
-    /// sending if it would exceed the model's context window. Off by default:
-    /// normal budgeting uses cheap local estimates, and this adds a network
-    /// round-trip per inference for providers with a remote count endpoint.
-    #[serde(default)]
-    pub exact_token_counting: bool,
-
-    /// Wall-clock timeout (seconds) for a Rhai script tool's `shell()` host call,
-    /// mirroring the built-in shell tool's own 60-second cap so a script can't
-    /// hang an agent on a runaway command. Defaults to `60`.
-    #[serde(default = "default_script_shell_timeout_secs")]
-    pub script_shell_timeout_secs: u64,
-
-    /// How long (seconds) a run may sit ready to work but unable to dispatch
-    /// before it is failed instead of left running.
-    ///
-    /// This only ever fires for something the runtime cannot resolve on its own -
-    /// today, a stage whose provider is not configured. Waiting for a busy
-    /// model's inference pool is ordinary backpressure and is never failed, no
-    /// matter how long it takes. Defaults to `60`; `0` disables the watchdog and
-    /// restores the old behaviour of waiting indefinitely.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_stall_timeout_secs")]
-    pub stall_timeout_secs: u64,
-
-    /// How many consecutive safety re-drives may find the tool lane full and no
-    /// run moving before the daemon widens the lane to break the jam.
-    ///
-    /// The daemon re-drives itself every 30 seconds, so the default of `10` is
-    /// five minutes of a full lane going nowhere. Relief only ever *adds*
-    /// capacity, never cancels anything, and is capped at one extra lane's worth
-    /// over the daemon's life, so it cannot run away.
-    ///
-    /// `0` turns relief off. Detection and reporting stay on either way, so
-    /// `lev ps` and the metrics still show the streak.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_dead_cycles_before_relief")]
-    pub dead_cycles_before_relief: u32,
-
-    /// How long (seconds) a run keeps its place in `lev ps` after the daemon
-    /// unloads it from memory.
-    ///
-    /// A terminal run used to leave the listing the moment it was unloaded,
-    /// which made a run that died on its first inference look exactly like a run
-    /// that had never been spawned. A scheduler polling the listing could only
-    /// tell the two apart with a stopwatch, and issue #205 is what that cost:
-    /// forty minutes of spawning work, timing out, and spawning it again.
-    ///
-    /// Defaults to `300`. `0` drops a run as soon as it finishes, which is the
-    /// old behaviour. The record lives in memory, so a restart clears it
-    /// whatever this is set to.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_finished_retention_secs")]
-    pub finished_retention_secs: u64,
-    /// How long (seconds) a per-agent MCP server may sit with zero live runs
-    /// leasing it before the daemon disconnects it (ending a stdio server's
-    /// child process). Long enough that back-to-back runs of a blueprint reuse
-    /// the warm connection; the next run that declares the server reconnects
-    /// lazily. `0` keeps every server connected for the daemon's life, which
-    /// was the old behaviour. Global `[[mcp_servers]]` from config.toml are
-    /// never disconnected regardless.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_mcp_idle_disconnect_secs")]
-    pub mcp_idle_disconnect_secs: u64,
-    /// How long (seconds) a run may sit in a state no part of the engine can
-    /// reach before it is failed instead of left reported as running.
-    ///
-    /// Not a general "this run looks slow" timeout, and never fires on one. An
-    /// agent waiting on the model, on a tool, on its sub-agents, or on a person
-    /// is holding the marker that says so, and is exempt however long it takes.
-    /// This only catches an agent holding *no* marker at all, which the engine's
-    /// own invariants say cannot happen and which nothing will ever look at
-    /// again. Such a run stays `running` in `meta.json` for the life of the
-    /// daemon and keeps whatever capacity an external scheduler assigned it,
-    /// which is issue #202.
-    ///
-    /// Defaults to `0`, which is off: this fails runs, and an upgrade that
-    /// starts killing work nobody asked it to kill is worse than the leak. `300`
-    /// is a reasonable value to set. Turning it on is also a way to find out
-    /// whether it is happening to you, since it says so in the log and in the
-    /// run's error.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_wedge_timeout_secs")]
-    pub wedge_timeout_secs: u64,
-    /// How many consecutive provider-fatal failures (out of credits, rejected
-    /// key) take a provider out of service for every run.
-    ///
-    /// Defaults to `3`. One 402 can just be a request asking for more output
-    /// tokens than the balance covers; three in a row is the account. While a
-    /// provider is out, runs move to their next candidate, and runs with none
-    /// left are failed by the stall watchdog rather than left "running".
-    ///
-    /// `0` disables the breaker, leaving per-run failover on its own.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_provider_failures_before_open")]
-    pub provider_failures_before_open: u32,
-
-    /// How long a provider stays out of service before one request is let
-    /// through to see whether it recovered.
-    ///
-    /// Defaults to `300` (five minutes). That probe either succeeds, which puts
-    /// the provider straight back into service, or fails and restarts the wait,
-    /// so topping up an account brings the factory back with no restart.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_provider_circuit_cooldown_secs")]
-    pub provider_circuit_cooldown_secs: u64,
-    /// How long (seconds) a prompt may go unanswered before the daemon resolves
-    /// it itself and lets the run carry on.
-    ///
-    /// Covers every prompt that waits on a person: an agent's `ask_user_*` /
-    /// `present_for_review` call, a tool-approval prompt, a taint gate, and a
-    /// blueprint interaction point. Before this existed, a run whose operator
-    /// had walked away sat in `WaitingInput` holding its slot until the daemon
-    /// restarted - hours, in the report that prompted it (issue #204).
-    ///
-    /// Expiry resolves the prompt exactly as cancelling it would: a tool
-    /// approval and a taint gate **deny**, an `ask_user_*` call is told nobody
-    /// answered, and an interaction point proceeds with no user text. Nothing is
-    /// approved on the strength of a timeout.
-    ///
-    /// Defaults to `3600` (one hour); `0` waits indefinitely.
-    ///
-    /// Read once at daemon start, so a change needs a daemon restart.
-    #[serde(default = "default_interaction_timeout_secs")]
-    pub interaction_timeout_secs: u64,
-
-    /// Most bytes one tool call may write to disk. Unset is unlimited.
-    ///
-    /// **Unset in code, set by `lev setup`.** How much an agent should write is
-    /// a judgement about what you are doing with it, so nothing is imposed on a
-    /// user who never opened this file - but a fresh install gets a concrete
-    /// number written here, where it is visible and can be deleted outright.
-    ///
-    /// The incident behind it (issue #252) was a single shell call appending in
-    /// a loop until the 60-second timeout: about 14 GB, from one call that
-    /// looked ordinary.
-    ///
-    /// A shell redirect is measured *after* the call, since the bytes go from
-    /// the shell to the file without passing through Leviath. So this stops the
-    /// call after the one that overran, not the one that did. `write_file` is
-    /// measured before, and is stopped outright.
-    ///
-    /// Running out of disk is checked separately and is never configurable: see
-    /// [`leviath_core::write_limits::MIN_FREE_BYTES`].
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_tool_call_write_bytes: Option<u64>,
-
-    /// Most bytes a whole run may write to disk. Unset is unlimited.
-    ///
-    /// The companion to `max_tool_call_write_bytes`, and the one that catches
-    /// what a per-call ceiling cannot: three calls of 12-14 GB each are
-    /// individually plausible and collectively a full disk. Same defaulting -
-    /// unset in code, written by `lev setup`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_run_write_bytes: Option<u64>,
-}
-
-impl LimitsConfig {
-    /// The write ceilings in effect, for the engine.
-    pub fn write_limits(&self) -> leviath_core::write_limits::WriteLimits {
-        leviath_core::write_limits::WriteLimits {
-            per_call: self.max_tool_call_write_bytes,
-            per_run: self.max_run_write_bytes,
-        }
-    }
-}
-
-impl Default for LimitsConfig {
-    fn default() -> Self {
-        Self {
-            max_concurrent_inferences: default_max_concurrent_inferences(),
-            max_concurrent_tools: default_max_concurrent_tools(),
-            default_max_iterations: default_default_max_iterations(),
-            exact_token_counting: false,
-            script_shell_timeout_secs: default_script_shell_timeout_secs(),
-            stall_timeout_secs: default_stall_timeout_secs(),
-            dead_cycles_before_relief: default_dead_cycles_before_relief(),
-            finished_retention_secs: default_finished_retention_secs(),
-            mcp_idle_disconnect_secs: default_mcp_idle_disconnect_secs(),
-            wedge_timeout_secs: default_wedge_timeout_secs(),
-            provider_failures_before_open: default_provider_failures_before_open(),
-            provider_circuit_cooldown_secs: default_provider_circuit_cooldown_secs(),
-            interaction_timeout_secs: default_interaction_timeout_secs(),
-            // Deliberately `None` here and concrete in `lev setup`: the code
-            // imposes no ceiling on a user who never opened the config, and a
-            // fresh install gets a number written where it can be seen and
-            // deleted. See the field docs.
-            max_tool_call_write_bytes: None,
-            max_run_write_bytes: None,
-        }
-    }
-}
-
-fn default_webhook_max_retries() -> u32 {
-    3
-}
-
-fn default_webhook_base_delay_ms() -> u64 {
-    500
-}
-
-fn default_webhook_max_delay_ms() -> u64 {
-    30_000
-}
-
-fn default_webhook_timeout_secs() -> u64 {
-    10
-}
-
-/// Completion-webhook delivery tuning.
-///
-/// A completion webhook is POSTed when a run reaches a terminal status. Delivery
-/// retries on transient failures (network errors, timeouts, 5xx, 429, 408) with
-/// exponential backoff. Each field has a safe default so `[webhook]` can be
-/// omitted entirely.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct WebhookConfig {
-    /// Number of retries **after** the first attempt (so total sends is
-    /// `max_retries + 1`). Defaults to `3`. Set `0` to disable retries.
-    #[serde(default = "default_webhook_max_retries")]
-    pub max_retries: u32,
-
-    /// Base backoff before the first retry, in milliseconds. Subsequent retries
-    /// double it (capped at `max_delay_ms`). Defaults to `500`.
-    #[serde(default = "default_webhook_base_delay_ms")]
-    pub base_delay_ms: u64,
-
-    /// Upper bound on any single backoff delay, in milliseconds. Defaults to
-    /// `30_000` (30s).
-    #[serde(default = "default_webhook_max_delay_ms")]
-    pub max_delay_ms: u64,
-
-    /// Per-attempt request timeout, in seconds. Defaults to `10`.
-    #[serde(default = "default_webhook_timeout_secs")]
-    pub timeout_secs: u64,
-}
-
-impl Default for WebhookConfig {
-    fn default() -> Self {
-        Self {
-            max_retries: default_webhook_max_retries(),
-            base_delay_ms: default_webhook_base_delay_ms(),
-            max_delay_ms: default_webhook_max_delay_ms(),
-            timeout_secs: default_webhook_timeout_secs(),
-        }
-    }
-}
-
-/// Provider configuration.
-///
-/// `Debug` is hand-written (see below) so the keys cannot be printed.
-#[derive(Clone, Default, Serialize, Deserialize)]
-pub struct ProviderConfig {
-    /// Anthropic API key
-    #[serde(default)]
-    pub anthropic_api_key: Option<String>,
-
-    /// OpenAI API key
-    #[serde(default)]
-    pub openai_api_key: Option<String>,
-
-    /// Google AI (Gemini) API key
-    #[serde(default)]
-    pub google_api_key: Option<String>,
-
-    /// Whether the Claude Code CLI transport is enabled.
-    ///
-    /// **Opt-in, and never selected for the user.** The CLI injects its own
-    /// context into every call - including the account email address on the
-    /// OAuth (subscription) path - which cannot be disabled. `lev setup` offers
-    /// it and defaults to declining, so a user who presses Enter through the
-    /// wizard ends up with it off.
-    #[serde(default)]
-    pub claude_code_enabled: bool,
-
-    /// Path to the `claude` executable. `None` resolves `claude` on `PATH`.
-    #[serde(default)]
-    pub claude_code_binary: Option<String>,
-
-    /// Reasoning effort for the Claude Code transport: `low` | `medium` |
-    /// `high` | `xhigh` | `max`.
-    ///
-    /// Always sent explicitly. Left to itself the CLI picks `high` with adaptive
-    /// thinking, spending output tokens and latency Leviath never asked for.
-    /// `None` uses [`leviath_providers::claude_code::DEFAULT_EFFORT`].
-    #[serde(default)]
-    pub claude_code_effort: Option<String>,
-
-    /// Host-wide failover chain, as `"provider/model"` entries, best first.
-    ///
-    /// Tried after a stage's own `models` list and the default model when the
-    /// provider in use stops answering (out of credits, rejected key). Entries
-    /// naming an unregistered provider are skipped, and a malformed entry is
-    /// ignored with a warning rather than failing the load.
-    ///
-    /// `provider/model` rather than a bare provider name because a failover
-    /// target needs a model to send; there is no sensible default per provider.
-    /// A blueprint that names one model has nowhere to go without this, which
-    /// is exactly how issue #201 took every agent down at once.
-    #[serde(default)]
-    pub fallback_order: Vec<String>,
-}
-
-/// Hand-written so the API keys can never be printed.
-///
-/// A `#[derive(Debug)]` here meant one `tracing::debug!(?config)` anywhere in
-/// the workspace - or one `dbg!`, or an `anyhow` context that formats a struct
-/// holding this - would put every provider key into the logs. Nothing did that
-/// today, which is exactly when it is cheap to foreclose: the type now cannot
-/// leak, so nobody has to remember not to.
-///
-/// Reports whether each key is *set*, which is what a debug line is actually
-/// asking, and mirrors the `RedactedConfig` the `/api/config` handler returns.
-impl std::fmt::Debug for ProviderConfig {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ProviderConfig")
-            .field("anthropic_api_key", &redacted(&self.anthropic_api_key))
-            .field("openai_api_key", &redacted(&self.openai_api_key))
-            .field("google_api_key", &redacted(&self.google_api_key))
-            .field("claude_code_enabled", &self.claude_code_enabled)
-            .field("claude_code_binary", &self.claude_code_binary)
-            .field("claude_code_effort", &self.claude_code_effort)
-            .field("fallback_order", &self.fallback_order)
-            .finish()
-    }
-}
-
-/// `"<set>"` or `"<unset>"` for an optional secret, for [`Debug`] output.
-fn redacted(value: &Option<String>) -> &'static str {
-    match value {
-        Some(_) => "<set>",
-        None => "<unset>",
-    }
-}
-
-/// Optional overrides for a Rhai script provider, from `[model_providers.<name>]`.
-///
-/// Every field is optional. Keys not recognized below flow into [`Self::extra`]
-/// and are forwarded to the script's `initialize(config)` alongside `base_url`
-/// and `api_key`.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ModelProviderConfig {
-    /// Script filename stem or path. Defaults to `<name>.rhai` in the providers
-    /// directory (`~/.leviath/providers/`).
-    #[serde(default)]
-    pub script: Option<String>,
-
-    /// API key forwarded to the script as `config.api_key` (a script may instead
-    /// read its own environment variable).
-    #[serde(default)]
-    pub api_key: Option<String>,
-
-    /// Base URL forwarded to the script as `config.base_url`.
-    #[serde(default)]
-    pub base_url: Option<String>,
-
-    /// Rate limit enforced by the Rust wrapper (requests/tokens per minute).
-    #[serde(default)]
-    pub rate_limit: Option<leviath_providers::RateLimitConfig>,
-
-    /// Any additional keys, forwarded verbatim into the script's `initialize`.
-    #[serde(flatten)]
-    pub extra: HashMap<String, toml::Value>,
 }
 
 impl Default for Config {
@@ -1566,6 +813,29 @@ fn set_dir_permissions_with(
     }
 }
 
+/// Serde default for a flag that ships on.
+///
+/// Shared by `[security]`, `[limits]` and `Config` itself, so it lives here
+/// rather than in whichever section happened to need it first: serde resolves
+/// a `default = "..."` path in the module the struct is defined in, so a helper
+/// three sections use has to be reachable from all three.
+pub(crate) fn default_true() -> bool {
+    true
+}
+
+/// The provider a config that names none is assumed to mean.
+///
+/// Exists so `default_provider` can carry `#[serde(default)]`: without one,
+/// every field on [`Config`] that lacked a default made a hand-written
+/// `config.toml` a parse error. Writing three lines to point Leviath at
+/// OpenRouter used to fail with `missing field `providers``, which names a
+/// table the user has no reason to know about and says nothing about what to
+/// add. Kept in sync with [`Config::default`] by
+/// `an_empty_config_file_parses_to_the_defaults`.
+pub(crate) fn default_provider_name() -> String {
+    "anthropic".to_string()
+}
+
 /// Serializes any test, anywhere in the crate, that mutates the process's
 /// current working directory (via `std::env::set_current_dir`) or whose
 /// assertions implicitly depend on it. Declared here (not inside `mod tests`)
@@ -1934,8 +1204,8 @@ mod dotenv_tests {
 mod tests {
     /// The published JSON Schema for `config.toml`, and a config exercising
     /// every section of it. Compiled in so neither can drift from what ships.
-    const CONFIG_SCHEMA: &str = include_str!("../../../docs/schema/config.schema.json");
-    const CONFIG_EXAMPLE: &str = include_str!("../../../docs/schema/config.example.toml");
+    const CONFIG_SCHEMA: &str = include_str!("../../../../docs/schema/config.schema.json");
+    const CONFIG_EXAMPLE: &str = include_str!("../../../../docs/schema/config.example.toml");
 
     /// Every way `value` fails `validator`. See the twin in `bundled.rs`.
     fn schema_problems(
